+++ web/src/interfaces/IMashStep.ts
@@ -6,3 +6,12 @@ export interface IMashStep {
   time: number;
   extendStepTimeIfNeeded: boolean;
 }
+
+export const defaultMashStep:IMashStep = {
+  index: 0,
+  name: 'New Step',
+  temperature: 0,
+  stepTime: 0,
+  time: 0,
+  extendStepTimeIfNeeded: true,
+};