+++ web/src/views/TempSettings.vue
@@ -1,6 +1,5 @@
 <script lang="ts" setup>
 import { ref, onMounted, onBeforeUnmount, inject } from 'vue';
-import { VDataTable } from 'vuetify/labs/VDataTable';
 import { mdiPencil, mdiDelete, mdiPalette } from '@mdi/js';
 import WebConn from '@/helpers/webConn';
 import { ITempSensor } from '@/interfaces/ITempSensor';
@@ -26,9 +25,9 @@ const dialog = ref<boolean>(false);
 const dialogDelete = ref<boolean>(false);
 
 const alert = ref<string>('');
-const alertType = ref<'error' | 'success' | 'warning' | 'info' >('info');
+const alertType = ref<'error' | 'success' | 'warning' | 'info'>('info');
 
-const defaultSensor:ITempSensor = {
+const defaultSensor: ITempSensor = {
   id: '',
   name: 'New Sensor',
   color: '#ffffff',
@@ -94,12 +93,12 @@ const closeDeleteDialog = async () => {
   dialogDelete.value = false;
 };
 
-const editItem = async (item:ITempSensor) => {
+const editItem = async (item: ITempSensor) => {
   editedItem.value = item;
   dialog.value = true;
 };
 
-const openDeleteDialog = async (item:ITempSensor) => {
+const openDeleteDialog = async (item: ITempSensor) => {
   editedItem.value = item;
   dialogDelete.value = true;
 };
@@ -128,14 +127,13 @@ const save = async () => {
 
 <template>
   <v-container class="pa-6" fluid>
-    <v-alert :type="alertType" v-if="alert" closable @click:close="alert = ''">{{alert}}</v-alert>
+    <v-alert :type="alertType" v-if="alert" closable @click:close="alert = ''">{{ alert }}</v-alert>
     <v-form fast-fail @submit.prevent>
       <v-data-table
         :headers="tableHeaders"
         :items="tempSensors"
         density="compact"
-        item-value="name"
-      >
+        item-value="name">
         <template v-slot:top>
           <v-toolbar density="compact">
             <v-toolbar-title>Temp Sensors</v-toolbar-title>
@@ -149,9 +147,9 @@ const save = async () => {
 
             <v-dialog v-model="dialog" max-width="500px">
               <v-card>
-                <v-card-title>
-                  <span class="text-h5">Edit</span>
-                </v-card-title>
+                <v-toolbar density="compact" color="dialog-header">
+                  <v-toolbar-title>Edit</v-toolbar-title>
+                </v-toolbar>
 
                 <v-card-text>
                   <v-container>
@@ -198,20 +196,20 @@ const save = async () => {
           </v-toolbar>
         </template>
         <template v-slot:[`item.actions`]="{ item }">
-          <v-icon size="small" class="me-2" @click="editItem(item.raw)" :icon="mdiPencil" />
-          <v-icon size="small" @click="openDeleteDialog(item.raw)" :icon="mdiDelete" />
+          <v-icon size="small" class="me-2" @click="editItem(item)" :icon="mdiPencil" />
+          <v-icon size="small" @click="openDeleteDialog(item)" :icon="mdiDelete" />
         </template>
         <template v-slot:[`item.useForControl`]="{ item }">
-          <v-checkbox-btn class="align-right justify-center" v-model="item.columns.useForControl" disabled />
+          <v-checkbox-btn class="align-right justify-center" v-model="item.useForControl" disabled />
         </template>
         <template v-slot:[`item.show`]="{ item }">
-          <v-checkbox-btn class="align-right justify-center" v-model="item.columns.show" disabled />
+          <v-checkbox-btn class="align-right justify-center" v-model="item.show" disabled />
         </template>
         <template v-slot:[`item.connected`]="{ item }">
-          <v-checkbox-btn class="align-right justify-center" v-model="item.columns.connected" disabled />
+          <v-checkbox-btn class="align-right justify-center" v-model="item.connected" disabled />
         </template>
         <template v-slot:[`item.color`]="{ item }">
-          <v-icon size="small" class="me-2" :icon="mdiPalette" :color="item.columns.color" />
+          <v-icon size="small" class="me-2" :icon="mdiPalette" :color="item.color" />
         </template>
       </v-data-table>